ELFRemove: also change DT_REL(A)COUNT when removing relocations
While we were already changing DT_REL(A)SZ to a smaller value
when removing relocations, we did not properly deal with the
(optional) DT_REL(A)COUNT tag. This leads to an assertion
when loading a tailored library with the DT_REL(A)COUNT tag
(such as glibc) as the loader will encounter a relocation
with a wrong type when processing R_XX_RELATIVE relocations
(because it expects DT_REL(A)COUNT of them while there will
actually be fewer after tailoring).

Fix it by also changing the DT_REL(A)COUNT tag if it exists.