ELFRemove: allow external debug directories
Allow external debug directories to be provided via
environment variables. These variables are the same as in
librarytrader, namely EXTERNAL_DEBUG_DIR and
EXTERNAL_BUILDID_DIR. As the filename does not consist of the
full path to the original file but already points to the
location of the tailored file, use the basename only. Note
that this requires either a flat directory with all debug
files or you need to provide all paths to the directories in
the EXTERNAL_DEBUG_DIR variable.