ELFRemove: workaround for buggy GNU loader behaviour
glibc had a bug in the calculation of relocation section
sizes before commit [0] (which went into version 2.23) which
leads to split-up relocation sections not working properly
when BIND_NOW is set.

Try to detect this situation by checking DT_FLAGS for
DF_BIND_NOW and the .note.ABI-tag section for compatibility
with a 2.6.32 kernel (which was upgraded to 3.2.0 around
version 2.26 for x86 and x86_64).

If both these criteria match, do not split up the relocation
sections by shortening .rela.dyn but instead repeat the last
remaining relocation as many times as needed to fill the gap
to .rela.plt.

In the future, we could also make this behaviour configurable
via an environment variable or something similar.

[0]: https://github.com/bminor/glibc/commit/fa19d5c48a6b36518ce68412e3bdde6bfa8aa4a6#diff-1d12cfc949b2b87b2ed291747dddc82495809c6a572bfa398ca7976385933291