ELFRemove: zero last bucket for GNU hash if it is empty
While the earlier buckets are correctly fixed up by the loop
iterating over the buckets and adapting the symbol indices,
the last bucket could point to an index larger than the last
symbol table index if all symbols in the last bucket were
removed. In the readelf output, this leads to a long tail of
bucket sizes in the bucket histogram.

Fix this by writing 0 to all buckets at the end of the list
in order to mark them as empty if no symbols remain in the
corresponding buckets.