ELFRemove: also zero out other bucket(s) if they become empty
So far, some buckets in a GNU hash table could point to
symbol indices which did not actually correspond to that
bucket if they became empty during the deletion process.

Improve the detection of those cases and zero out entries
directly in the buckets array if the buckets become empty.