ELFRemove: recreate SysV hash table instead of modifying it
Removing entries from the old style hash tables is more
complicated than rebuilding it 'from scratch' as we need to
modify entries multiple times while iterating over the
removed symbols.

As the construction of the hash table is quite simple, we
can rather build the table from the finalized .dynsym table
after the symbols have been removed. In the process, we can
also shrink the number of buckets if the number of symbols
in .dynsym falls into a smaller range (the options for
hash table sizes are taken from the gold linker/binutils).

This makes removal of 12000 out of 36000 functions from
libLLVM-7.so.1 finish in 6 seconds instead of 80 seconds.