ELFRemove: fix up sh_info values after removing symbols
The ELF specification states that SHT_DYNSYM and SHT_SYMTAB
section header table entries contain the index of the first
non-local symbol in the sh_info field. libbfd actually checks
this when parsing the ELF file so an invalid value (e.g. the
value is bigger than sh_size * sh_entsize) will lead to an
'unrecognized file format' error in tools like objdump or
strip.

Let's fix the value just as we did with the sh_size member as
well. While we're at it, make the member fixup code a bit more
generic and reusable.