ELFRemove, remove_tool_libtrader: strip debug info before modifying anything
As stripping happened before any modification to the .dynsym table
(and thus, all associated other tables) but _after_ removing local
functions, 'strip' could already change the layout of the output
file as removing local functions also changes the relocation table.

When enough local functions with relocations are removed, the
relocation section could become so small that a whole page fits
between the end of the relocation section and the following section
(mostly .text or .init). In this case, strip would already move
sections forward in the output file and generate a mismatch between
the keep files for shrinkelf and the actual file.

We fix this by stripping the files at the very beginning of the
process - before any modifications to the file - but strip any
present debug information into a seperate file and introducing this
information into the process via the already existing environment
variables. Additionally, change the detection logic for .symtab
slightly to allow files without a .gnu.build-id section but
containing a .symtab section to be detected as valid debugging
information files.

In the future, we might need to add some more checking if the file
matches the ELF file (e.g., by matching up the section header tables
if no .gnu.build-id was found).