remove_tool_libtrader.py: start timing after ELFRemove creation
With the newly added debuginfod support in the last commit, the
construction of the ELFRemove object might take quite long if
the corresponding debug information file needs to be downloaded.

Let's exclude the initial creation of the object from our timing
measurements as nothing is actually removed at this stage.