ELFRemove: fix relocation removal
With some combinations of _RELATIVE and _64 relocations for the same
symbol, we could miss removal and thus fail to fix up other
relocations properly.

Fix this by skipping the increment if we're allowed to push
relocations together as we end up at the correct location after
the call to bisect.bisect_left() anyways.