ELFRemove: keep relocations continuous if we're processing ld-linux
Before glibc commit f64f4ce06930 ("elf: Assume disjointed .rela.dyn
and .rela.plt for loader"), .rela.dyn and .rela.plt sections had to
be continuous for ld-linux itself. Note that this is similar to the
description in commit 6822a43c5923 where this was fixed for BIND_NOW
binaries and loaders pre-2014.

For now, let's always keep relocations continuous if we encounter
the ld-linux-* binary.