ELFRemove: compute number of executable bytes via sections, not segments
If LOAD statements are considered, we overestimate the number of
'actual' instructions as the initial LOAD statement often also
includes the ELF metadata at the beginning of the file. If we have
a file with a large symbol table, for example, all these bytes
are counted towards the number of instruction bytes as the
containing mapping will be R-X. If we're only comparing bytes of
functions, however, we should look at the sections with executable
code only, so check the section header flags instead of segments.