ELFRemove: shrink .dynstr section
When symbols are removed from the file, their associated metadata was
removed from the symbol table but their name was still present as a
string in the dynamic string table. Let's make this section more
compact as well to save even more space.

Note that this required some reworking of dynamic tag setting to work
as we have to modify multiple tags of the same type (DT_NEEDED) so
the relevant functions have been altered to accommodate this.