ELFRemove, remove_tool_libtrader: allow removal of SHN_UNDEF entries from .dynsym
With this change, we can remove unused dependencies based on the
call graph information from librarytrader and possibly shrink the
resulting files even more.

Previously, we had to mark every undefined function in .dynsym as
used as the dynamic loader would otherwise fail to locate the target
function in the tailored libraries, even if the function was never
actually used. With this change, we can also remove SHN_UNDEF entries
which are never referenced from any local or exported function which
leads to more potential of removal in 'lower' libraries. Note that
this only makes a difference if you are using a version of
librarytrader with the TAILOR_BINARIES commit applied and the
`--leave-undef-unused` parameter for `running_analysis.py` has been
used.