ELFRemove: fix for SHN_UNDEF removal from symbol table
The imported symbols do not need to be contiguous at the beginning
of the symbol table but can also occur later. However, the
symoffset parameter in the GNU hash table only accounts for the
first contiguous block. Fix it by not only splitting undefined and
defined symbols from the removal list but keeping undefined symbols
which come after .gnu.hash['symoffset'] as part of the defined
symbols list.