ELFRemove: speed up .dynstr shrinking
Instead of removing items from a list (which is O(n)) let's
count the number of references to indices in the dynamic
string table instead. Note that we can not simply use a set
as a string can be referenced multiple times (for example,
if multiple versions of the same function name or multiple
version table entries pointing to version or file names
exist). Instead, use the string ranges as indices in a
dictionary and increase/decrease the value when we find a
reference.